Add Umami analytics script to site header
The commit includes the Umami analytics script to the head of the site. The script will allow access to more detailed analytics data for the website contributing to more informed decision-making in website design and development. The script is inserted with the "defer" attribute to prevent it from blocking the parsing of the HTML while it's loading.