Add package-lock.json with project dependencies
This commit introduces a package-lock.json file, which is used to lock down the versions of a project's dependencies. This means all developers working on the project will use the same exact versions of all dependencies (such as "@11ty/eleventy", "@tailwindcss/forms", and "autoprefixer"), mitigating the potential for bugs caused by version differences.