Remove edit-post script and modify new-post script
The edit-post.js script was removed because it was unnecessary. The new-post.js script has been updated to replace 'year' with 'date' in the file path structure, and the 'date' attribute has been removed from the script. The package.json file has also been updated accordingly to reflect the removal of the edit-post command.