Update several package versions
Several packages in the package.json and package-lock.json files have been updated, including "inquirer", "isbot", "postcss", and "tailwind-config-viewer". This update brings these packages up to their latest versions and ensures better stability and performance for the project as a whole. The removal of "@ljharb/through" from "inquirer" dependencies is also included in this update.